Move pointer down encoding to interaction encoders
